+++ R/fdadbscan.R
@@ -47,6 +47,8 @@ fdadbscan <- function(x, y,
                       check_total_dissimilarity = TRUE,
                       compute_overall_center = FALSE) {
   call <- rlang::call_match(defaults = TRUE)
+  callname <- rlang::call_name(call)
+  callargs <- rlang::call_args(call)
 
   l <- format_inputs(x, y)
   x <- l$x
@@ -66,6 +68,9 @@ fdadbscan <- function(x, y,
 
   if (warping_class == "none" && cluster_on_phase)
     cli::cli_abort("It makes no sense to cluster based on phase variability if no alignment is performed.")
+
+  callargs$warping_class <- warping_class
+  callargs$metric <- metric
 
   if (use_verbose)
     cli::cli_alert_info("Computing the distance matrix...")
@@ -128,41 +133,31 @@ fdadbscan <- function(x, y,
   if (use_verbose)
     cli::cli_alert_info("Consolidating output...")
 
-  # grids <- purrr::map(kmresults, \(km) km$grids[1, ])
-  grids <- purrr::imap(labels, \(.x, .y) {
-    if (.x == 0)
-      x[.y, ]
-    else
-      kmresults[[.x]]$grids[1, ]
-  })
-  grids <- do.call(rbind, grids)
   original_curves <- array(dim = c(N, L, M))
-  aligned_curves <- array(dim = c(N, L, M))
+  original_curves[labels == 0, , ] <- y[labels == 0, , ]
+  original_grids <- matrix(nrow = N, ncol = M)
+  original_grids[labels == 0, ] <- x[labels == 0, ]
+  aligned_grids <- matrix(nrow = N, ncol = M)
+  aligned_grids[labels == 0, ] <- x[labels == 0, ]
   center_curves <- array(dim = c(n_clusters, L, M))
-  warpings <- matrix(nrow = N, ncol = M)
+  center_grids <- matrix(nrow = n_clusters, ncol = M)
   dtc <- numeric(N)
+  dtc[labels == 0] <- 0
   for (k in 1:n_clusters) {
     cluster_ids <- which(labels == k)
     original_curves[cluster_ids, , ] <- kmresults[[k]]$original_curves
-    aligned_curves[cluster_ids, , ] <- kmresults[[k]]$aligned_curves
+    original_grids[cluster_ids, ] <- kmresults[[k]]$original_grids
+    aligned_grids[cluster_ids, ] <- kmresults[[k]]$aligned_grids
     center_curves[k, , ] <- kmresults[[k]]$center_curves
-    warpings[cluster_ids, ] <- kmresults[[k]]$warpings
+    center_grids[k, ] <- kmresults[[k]]$center_grids[1, ]
     dtc[cluster_ids] <- kmresults[[k]]$distances_to_center
-  }
-  noise_ids <- which(labels == 0)
-  if (length(noise_ids) > 0) {
-    original_curves[noise_ids, , ] <- y[noise_ids, , ]
-    aligned_curves[noise_ids, , ] <- y[noise_ids, , ]
-    warpings[noise_ids, ] <- x[noise_ids, ]
-    dtc[noise_ids] <- 0
   }
 
   silhouettes <- NULL
   if (n_clusters > 1) {
     D <- fdadist(
-      # x = grids[labels, ],
-      x = grids,
-      y = aligned_curves,
+      x = aligned_grids,
+      y = original_curves,
       warping_class = "none",
       metric = metric
     )
@@ -171,10 +166,10 @@ fdadbscan <- function(x, y,
 
   out <- list(
     original_curves = original_curves,
-    aligned_curves = aligned_curves,
+    original_grids = original_grids,
+    aligned_grids = aligned_grids,
     center_curves = center_curves,
-    warpings = warpings,
-    grids = grids,
+    center_grids = center_grids,
     n_clusters = n_clusters,
     memberships = labels,
     distances_to_center = dtc,
@@ -182,8 +177,8 @@ fdadbscan <- function(x, y,
     amplitude_variation = sum(purrr::map_dbl(kmresults, "amplitude_variation")),
     total_variation = sum(purrr::map_dbl(kmresults, "total_variation")),
     n_iterations = 0,
-    call_name = rlang::call_name(call),
-    call_args = rlang::call_args(call)
+    call_name = callname,
+    call_args = callargs
   )
 
   as_caps(out)