+++ R/fdahclust.R
@@ -60,6 +60,8 @@ fdahclust <- function(x, y = NULL,
                       check_total_dissimilarity = TRUE,
                       compute_overall_center = FALSE) {
   call <- rlang::call_match(defaults = TRUE)
+  callname <- rlang::call_name(call)
+  callargs <- rlang::call_args(call)
 
   l <- format_inputs(x, y)
   x <- l$x
@@ -80,6 +82,10 @@ fdahclust <- function(x, y = NULL,
 
   if (warping_class == "none" && cluster_on_phase)
     cli::cli_abort("It makes no sense to cluster based on phase variability if no alignment is performed.")
+
+  callargs$warping_class <- warping_class
+  callargs$metric <- metric
+  callargs$linkage_criterion <- linkage_criterion
 
   if (use_verbose)
     cli::cli_alert_info("Computing the distance matrix...")
@@ -130,27 +136,27 @@ fdahclust <- function(x, y = NULL,
   if (use_verbose)
     cli::cli_alert_info("Consolidating output...")
 
-  grids <- purrr::map(kmresults, \(km) km$grids[1, ])
-  grids <- do.call(rbind, grids)
   original_curves <- array(dim = c(N, L, M))
-  aligned_curves <- array(dim = c(N, L, M))
+  original_grids <- matrix(nrow = N, ncol = M)
+  aligned_grids <- matrix(nrow = N, ncol = M)
   center_curves <- array(dim = c(n_clusters, L, M))
-  warpings <- matrix(nrow = N, ncol = M)
+  center_grids <- matrix(nrow = n_clusters, ncol = M)
   dtc <- numeric(N)
   for (k in 1:n_clusters) {
     cluster_ids <- which(labels == k)
     original_curves[cluster_ids, , ] <- kmresults[[k]]$original_curves
-    aligned_curves[cluster_ids, , ] <- kmresults[[k]]$aligned_curves
+    original_grids[cluster_ids, ] <- kmresults[[k]]$original_grids
+    aligned_grids[cluster_ids, ] <- kmresults[[k]]$aligned_grids
     center_curves[k, , ] <- kmresults[[k]]$center_curves
-    warpings[cluster_ids, ] <- kmresults[[k]]$warpings
+    center_grids[k, ] <- kmresults[[k]]$center_grids[1, ]
     dtc[cluster_ids] <- kmresults[[k]]$distances_to_center
   }
 
   silhouettes <- NULL
   if (n_clusters > 1) {
     D <- fdadist(
-      x = grids[labels, ],
-      y = aligned_curves,
+      x = aligned_grids,
+      y = original_curves,
       warping_class = "none",
       metric = metric
     )
@@ -159,10 +165,10 @@ fdahclust <- function(x, y = NULL,
 
   out <- list(
     original_curves = original_curves,
-    aligned_curves = aligned_curves,
+    original_grids = original_grids,
+    aligned_grids = aligned_grids,
     center_curves = center_curves,
-    warpings = warpings,
-    grids = grids,
+    center_grids = center_grids,
     n_clusters = n_clusters,
     memberships = labels,
     distances_to_center = dtc,
@@ -170,8 +176,8 @@ fdahclust <- function(x, y = NULL,
     amplitude_variation = sum(purrr::map_dbl(kmresults, "amplitude_variation")),
     total_variation = sum(purrr::map_dbl(kmresults, "total_variation")),
     n_iterations = 0,
-    call_name = rlang::call_name(call),
-    call_args = rlang::call_args(call)
+    call_name = callname,
+    call_args = callargs
   )
 
   as_caps(out)