+++ tests/testthat/test-fdakmeans.R
@@ -23,20 +23,20 @@ test_that('`fdakmeans()` works with fda::fd input object.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -62,20 +62,20 @@ test_that('`fdakmeans()` works with funData::funData input object.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -101,20 +101,20 @@ test_that('`fdakmeans()` works with fixed initial seeds.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -142,20 +142,20 @@ test_that('`fdakmeans()` works with kmeans++ seeding strategy.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -189,20 +189,20 @@ test_that('`fdakmeans()` works with exhaustive-kmeans++ seeding strategy.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -234,21 +234,21 @@ test_that('`fdakmeans()` works with exhaustive seeding strategy.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
-  expect_equal(out$n_iterations, 3)
+  expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
 })
@@ -273,20 +273,20 @@ test_that('`fdakmeans()` works with hclust seeding strategy.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -312,20 +312,20 @@ test_that('`fdakmeans()` works with dilation warping.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, 2)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -351,20 +351,20 @@ test_that('`fdakmeans()` works with no warping.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 4)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -390,20 +390,20 @@ test_that('`fdakmeans()` works with shift warping.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -429,21 +429,60 @@ test_that('`fdakmeans()` works with srsf warping.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 1)
+  expect_equal(out$call_name, "fdakmeans")
+  expect_true(inherits(out$call_args, "list"))
+})
+
+test_that('`fdakmeans()` works with median centroid.', {
+  dims <- dim(simulated30_sub$y)
+  N <- dims[1]
+  L <- dims[2]
+  P <- dims[3]
+  K <- 2L
+
+  out <- fdakmeans(
+    simulated30_sub$x,
+    simulated30_sub$y,
+    seeds = c(1, 21),
+    n_clusters = K,
+    centroid_type = "median",
+    warping_class = "affine",
+    metric = "pearson",
+    use_verbose = FALSE
+  )
+
+  expect_true(is_caps(out))
+  expect_equal(length(out), 14)
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
+                      "amplitude_variation", "total_variation", "n_iterations",
+                      "call_name", "call_args")
+  expect_equal(names(out), expected_names)
+  expect_equal(dim(out$original_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
+  expect_equal(dim(out$center_curves), c(K, L, P))
+  expect_equal(dim(out$center_grids), c(K, P))
+  expect_equal(out$n_clusters, K)
+  expect_equal(length(out$memberships), N)
+  expect_equal(length(out$distances_to_center), N)
+  expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
 })
@@ -468,20 +507,20 @@ test_that('`fdakmeans()` works with medoid centroid.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -507,20 +546,20 @@ test_that('`fdakmeans()` works with lowess centroid.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -546,21 +585,21 @@ test_that('`fdakmeans()` works with poly centroid.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
-  expect_equal(out$n_iterations, 1)
+  expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
 })
@@ -585,20 +624,20 @@ test_that('`fdakmeans()` works with l2 metric.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -625,20 +664,20 @@ test_that('`fdakmeans()` works when clustering on phase.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -664,20 +703,20 @@ test_that('`fdakmeans()` works in verbose mode.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -704,20 +743,20 @@ test_that('`fdakmeans()` works with parallel mode on distance calculation.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -744,20 +783,20 @@ test_that('`fdakmeans()` works with fence adaptive algorithm.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))
@@ -784,20 +823,20 @@ test_that('`fdakmeans()` works with computation of overall center.', {
 
   expect_true(is_caps(out))
   expect_equal(length(out), 14)
-  expected_names <- c("original_curves", "aligned_curves", "center_curves",
-                      "warpings", "grids", "n_clusters", "memberships",
-                      "distances_to_center", "silhouettes",
+  expected_names <- c("original_curves", "original_grids", "aligned_grids",
+                      "center_curves", "center_grids", "n_clusters",
+                      "memberships", "distances_to_center", "silhouettes",
                       "amplitude_variation", "total_variation", "n_iterations",
                       "call_name", "call_args")
   expect_equal(names(out), expected_names)
   expect_equal(dim(out$original_curves), dims)
-  expect_equal(dim(out$aligned_curves), dims)
+  expect_equal(dim(out$original_grids), c(N, P))
+  expect_equal(dim(out$aligned_grids), c(N, P))
   expect_equal(dim(out$center_curves), c(K, L, P))
-  expect_equal(dim(out$grids), c(K, P))
+  expect_equal(dim(out$center_grids), c(K, P))
   expect_equal(out$n_clusters, K)
   expect_equal(length(out$memberships), N)
   expect_equal(length(out$distances_to_center), N)
-  expect_equal(dim(out$warpings), c(N, P))
   expect_equal(out$n_iterations, 2)
   expect_equal(out$call_name, "fdakmeans")
   expect_true(inherits(out$call_args, "list"))